Tests/requirements.txt
pytest==8.3.5
httpx==0.28.1
fastapi==0.115.11
pillow==11.0.0